Start linking to a live bingo backend. Connecting to a local version of `spyrothon/bingo-league` to load the bingo board and eventually receive updates on the state of the board.
Also adds a `dashboards` folder for adding management dashboards in the near future. These will manage the data on the screen, primarily the team names and other statistics. The bingo board itself will be updated by players through the `bingo-league` app, but the moderator will have a dashboard to override any updates for the stream.